refactor layout of parser
