+++ cmd/parse_ratengine_data_ghc/main.go
@@ -4,7 +4,6 @@ import (
 	"flag"
 	"fmt"
 	"log"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -327,363 +326,5 @@ func process(params paramConfig, sheetIndex int) error {
 
 	// Verification and Process completed
 	log.Printf("Completed processing sheet index %d with description %s\n", sheetIndex, description)
-	return nil
-}
-
-/*************************************************************************/
-// Shared Helper functions
-/*************************************************************************/
-
-// A safe way to get a cell from a slice of cells, returning empty string if not found
-func getCell(cells []*xlsx.Cell, i int) string {
-	if len(cells) > i {
-		return cells[i].String()
-	}
-
-	return ""
-}
-
-// Gotta have a stringPointer function. Returns nil if empty string
-func stringPointer(s string) *string {
-	if s == "" {
-		return nil
-	}
-
-	return &s
-}
-
-func getInt(from string) int {
-	i, err := strconv.Atoi(from)
-	if err != nil {
-		if strings.HasSuffix(err.Error(), ": invalid syntax") {
-			f, ferr := strconv.ParseFloat(from, 32)
-			if ferr != nil {
-				return 0
-			}
-			if f != 0.0 {
-				return int(f)
-			}
-		}
-		log.Fatalf("ERROR: getInt() Atoi & ParseFloat failed to convert <%s> error %s, returning 0\n", from, err.Error())
-	}
-
-	return i
-}
-
-func checkError(message string, err error) {
-	if err != nil {
-		log.Fatal(message, err)
-	}
-}
-
-func removeFirstDollarSign(s string) string {
-	return strings.Replace(s, "$", "", 1)
-}
-
-func removeWhiteSpace(stripString string) string {
-	space := regexp.MustCompile(`\s`)
-	s := space.ReplaceAllString(stripString, "")
-
-	return s
-}
-
-func createCsvWriter(create bool, sheetIndex int, runTime time.Time) *createCsvHelper {
-	var createCsv createCsvHelper
-
-	if create == true {
-		err := createCsv.createCsvWriter(xlsxDataSheets[sheetIndex].generateOutputFilename(sheetIndex, runTime))
-		checkError("Failed to create CSV writer", err)
-	} else {
-		return nil
-	}
-	return &createCsv
-}
-
-/*************************************************************************/
-// GHC Rate Engine XLSX Verification and Process functions
-/*************************************************************************/
-
-// verifyDomesticLinehaulPrices: verification for 2a) Domestic Linehaul Prices
-var verifyDomesticLinehaulPrices verifyXlsxSheet = func(params paramConfig, sheetIndex int) error {
-
-	if dLhWeightBandNumCells != dLhWeightBandNumCellsExpected {
-		return fmt.Errorf("parseDomesticLinehaulPrices(): Exepected %d columns per weight band, found %d defined in golang parser", dLhWeightBandNumCellsExpected, dLhWeightBandNumCells)
-	}
-
-	if len(dLhWeightBands) != dLhWeightBandCountExpected {
-		return fmt.Errorf("parseDomesticLinehaulPrices(): Exepected %d weight bands, found %d defined in golang parser", dLhWeightBandCountExpected, len(dLhWeightBands))
-	}
-
-	// XLSX Sheet consts
-	const xlsxDataSheetNum int = 6  // 2a) Domestic Linehaul Prices
-	const feeColIndexStart int = 6  // start at column 6 to get the rates
-	const feeRowIndexStart int = 14 // start at row 14 to get the rates
-	const serviceAreaNumberColumn int = 2
-	const originServiceAreaColumn int = 3
-	const serviceScheduleColumn int = 4
-	const numEscalationYearsToProcess int = 2
-
-	// Check headers
-	const feeRowMilageHeaderIndexStart int = (feeRowIndexStart - 3)
-	const verifyHeaderIndexEnd int = (feeRowMilageHeaderIndexStart + 2)
-
-	if xlsxDataSheetNum != sheetIndex {
-		return fmt.Errorf("verifyDomesticLinehaulPrices expected to process sheet %d, but received sheetIndex %d", xlsxDataSheetNum, sheetIndex)
-	}
-
-	dataRows := params.xlsxFile.Sheets[xlsxDataSheetNum].Rows[feeRowMilageHeaderIndexStart:verifyHeaderIndexEnd]
-	for dataRowsIndex, row := range dataRows {
-		colIndex := feeColIndexStart
-		// For number of baseline + Escalation years
-		for escalation := 0; escalation < numEscalationYearsToProcess; escalation++ {
-			// For each Rate Season
-			for _, r := range rateSeasons {
-				// For each weight band
-				for _, w := range dLhWeightBands {
-					// For each milage range
-					for dLhMilesRangesIndex, m := range dLhMilesRanges {
-						// skip the last index because the text is not easily checked
-						if dLhMilesRangesIndex == len(dLhMilesRanges)-1 {
-							colIndex++
-							continue
-						}
-						verificationLog := fmt.Sprintf(" , verfication for row index: %d, colIndex: %d, Escalation: %d, rateSeasons %v, dLhWeightBands %v",
-							dataRowsIndex, colIndex, escalation, r, w)
-						if dataRowsIndex == 0 {
-							if m.lower != getInt(getCell(row.Cells, colIndex)) {
-								return fmt.Errorf("format error: From Miles --> does not match expected number expected %d got %s\n%s", m.lower, getCell(row.Cells, colIndex), verificationLog)
-							}
-							if "ServiceAreaNumber" != removeWhiteSpace(getCell(row.Cells, serviceAreaNumberColumn)) {
-								return fmt.Errorf("format error: Header <ServiceAreaNumber> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, serviceAreaNumberColumn)), verificationLog)
-							}
-							if "OriginServiceArea" != removeWhiteSpace(getCell(row.Cells, originServiceAreaColumn)) {
-								return fmt.Errorf("format error: Header <OriginServiceArea> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, originServiceAreaColumn)), verificationLog)
-							}
-							if "ServicesSchedule" != removeWhiteSpace(getCell(row.Cells, serviceScheduleColumn)) {
-								return fmt.Errorf("format error: Header <SServicesSchedule> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, serviceScheduleColumn)), verificationLog)
-							}
-						} else if dataRowsIndex == 1 {
-							if m.upper != getInt(getCell(row.Cells, colIndex)) {
-								return fmt.Errorf("format error: To Miles --> does not match expected number expected %d got %s\n%s", m.upper, getCell(row.Cells, colIndex), verificationLog)
-							}
-						} else if dataRowsIndex == 2 {
-							if "EXAMPLE" != getCell(row.Cells, originServiceAreaColumn) {
-								return fmt.Errorf("format error: Filler text <EXAMPLE> is missing got <%s> instead\n%s", getCell(row.Cells, originServiceAreaColumn), verificationLog)
-							}
-						}
-						colIndex++
-					}
-				}
-				colIndex++ // skip 1 column (empty column) before starting next Rate type
-			}
-		}
-	}
-
-	return nil
-}
-
-// parseDomesticLinehaulPrices: parser for 2a) Domestic Linehaul Prices
-var parseDomesticLinehaulPrices processXlsxSheet = func(params paramConfig, sheetIndex int) error {
-	// Create CSV writer to save data to CSV file, returns nil if params.saveToFile=false
-	csvWriter := createCsvWriter(params.saveToFile, sheetIndex, params.runTime)
-	if csvWriter != nil {
-		defer csvWriter.close()
-
-		// Write header to CSV
-		dp := domesticLineHaulPrice{}
-		csvWriter.write(dp.csvHeader())
-	}
-
-	// XLSX Sheet consts
-	const xlsxDataSheetNum int = 6  // 2a) Domestic Linehaul Prices
-	const feeColIndexStart int = 6  // start at column 6 to get the rates
-	const feeRowIndexStart int = 14 // start at row 14 to get the rates
-	const serviceAreaNumberColumn int = 2
-	const originServiceAreaColumn int = 3
-	const serviceScheduleColumn int = 4
-	const numEscalationYearsToProcess int = sharedNumEscalationYearsToProcess
-
-	if xlsxDataSheetNum != sheetIndex {
-		return fmt.Errorf("parseDomesticLinehaulPrices expected to process sheet %d, but received sheetIndex %d", xlsxDataSheetNum, sheetIndex)
-	}
-
-	dataRows := params.xlsxFile.Sheets[xlsxDataSheetNum].Rows[feeRowIndexStart:]
-	for _, row := range dataRows {
-		colIndex := feeColIndexStart
-		// For number of baseline + Escalation years
-		for escalation := 0; escalation < numEscalationYearsToProcess; escalation++ {
-			// For each Rate Season
-			for _, r := range rateSeasons {
-				// For each weight band
-				for _, w := range dLhWeightBands {
-					// For each milage range
-					for _, m := range dLhMilesRanges {
-						domPrice := domesticLineHaulPrice{
-							ServiceAreaNumber: getInt(getCell(row.Cells, serviceAreaNumberColumn)),
-							OriginServiceArea: getCell(row.Cells, originServiceAreaColumn),
-							ServiceSchedule:   getInt(getCell(row.Cells, serviceScheduleColumn)),
-							Season:            r,
-							WeightBand:        w,
-							MilesRange:        m,
-							Escalation:        escalation,
-							Rate:              getCell(row.Cells, colIndex),
-						}
-						colIndex++
-						if params.showOutput == true {
-							log.Println(domPrice.toSlice())
-						}
-						if csvWriter != nil {
-							csvWriter.write(domPrice.toSlice())
-						}
-					}
-				}
-				colIndex++ // skip 1 column (empty column) before starting next Rate type
-			}
-		}
-	}
-
-	return nil
-}
-
-// verifyDomesticServiceAreaPrices: verification 2b) Dom. Service Area Prices
-var verifyDomesticServiceAreaPrices verifyXlsxSheet = func(params paramConfig, sheetIndex int) error {
-	// XLSX Sheet consts
-	const xlsxDataSheetNum int = 7  // 2a) Domestic Linehaul Prices
-	const feeColIndexStart int = 6  // start at column 6 to get the rates
-	const feeRowIndexStart int = 10 // start at row 10 to get the rates
-	const serviceAreaNumberColumn int = 2
-	const serviceAreaNameColumn int = 3
-	const serviceScheduleColumn int = 4
-	const sITPickupDeliveryScheduleColumn int = 5
-	const numEscalationYearsToProcess int = 4
-
-	// Check headers
-	const feeRowMilageHeaderIndexStart int = (feeRowIndexStart - 2)
-	const verifyHeaderIndexEnd int = (feeRowMilageHeaderIndexStart + 2)
-
-	if xlsxDataSheetNum != sheetIndex {
-		return fmt.Errorf("verifyDomesticServiceAreaPrices expected to process sheet %d, but received sheetIndex %d", xlsxDataSheetNum, sheetIndex)
-	}
-
-	// Verify header strings
-	repeatingHeaders := []string{
-		"Shorthaul Price (per cwt per mile)",
-		"Origin / Destination Price (per cwt)",
-		"Origin Pack Price (per cwt)",
-		"Destination Unpack Price (per cwt)",
-		"Origin / Destination SIT First Day & Warehouse Handling (per cwt)",
-		"Origin / Destination SIT Add'l Days (per cwt)",
-		"SIT Pickup / Delivery ≤50 miles (per cwt)",
-	}
-
-	dataRows := params.xlsxFile.Sheets[xlsxDataSheetNum].Rows[feeRowMilageHeaderIndexStart:verifyHeaderIndexEnd]
-	for dataRowsIndex, row := range dataRows {
-		colIndex := feeColIndexStart
-		// For number of baseline + Escalation years
-		for escalation := 0; escalation < numEscalationYearsToProcess; escalation++ {
-			// For each Rate Season
-			for _, r := range rateSeasons {
-				verificationLog := fmt.Sprintf(" , verfication for row index: %d, colIndex: %d, Escalation: %d, rateSeasons %v",
-					dataRowsIndex, colIndex, escalation, r)
-
-				if dataRowsIndex == 0 {
-					if "ServiceAreaNumber" != removeWhiteSpace(getCell(row.Cells, serviceAreaNumberColumn)) {
-						return fmt.Errorf("format error: Header <ServiceAreaNumber> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, serviceAreaNumberColumn)), verificationLog)
-					}
-					if "ServiceAreaName" != removeWhiteSpace(getCell(row.Cells, serviceAreaNameColumn)) {
-						return fmt.Errorf("format error: Header <ServiceAreaName> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, serviceAreaNameColumn)), verificationLog)
-					}
-					if "ServicesSchedule" != removeWhiteSpace(getCell(row.Cells, serviceScheduleColumn)) {
-						return fmt.Errorf("format error: Header <ServicesSchedule> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, serviceScheduleColumn)), verificationLog)
-					}
-
-					if "SITPickup/DeliverySchedule" != removeWhiteSpace(getCell(row.Cells, sITPickupDeliveryScheduleColumn)) {
-						return fmt.Errorf("format error: Header <SIT Pickup / Delivery> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, sITPickupDeliveryScheduleColumn)), verificationLog)
-					}
-
-					for _, repeatingHeader := range repeatingHeaders {
-						if removeWhiteSpace(repeatingHeader) != removeWhiteSpace(getCell(row.Cells, colIndex)) {
-							return fmt.Errorf("format error: Header contains <%s> is missing got <%s> instead\n%s", removeWhiteSpace(repeatingHeader), removeWhiteSpace(getCell(row.Cells, colIndex)), verificationLog)
-						}
-						colIndex++
-					}
-					colIndex++ // skip 1 column (empty column) before starting next Rate type
-				} else if dataRowsIndex == 1 {
-					if "EXAMPLE" != removeWhiteSpace(getCell(row.Cells, serviceAreaNameColumn)) {
-						return fmt.Errorf("format error: Filler text <EXAMPLE> is missing got <%s> instead\n%s", removeWhiteSpace(getCell(row.Cells, serviceAreaNameColumn)), verificationLog)
-					}
-				}
-
-			}
-
-		}
-	}
-	return nil
-}
-
-// parseDomesticServiceAreaPrices: parser for: 2b) Dom. Service Area Prices
-var parseDomesticServiceAreaPrices processXlsxSheet = func(params paramConfig, sheetIndex int) error {
-	// Create CSV writer to save data to CSV file, returns nil if params.saveToFile=false
-	csvWriter := createCsvWriter(params.saveToFile, sheetIndex, params.runTime)
-	if csvWriter != nil {
-		defer csvWriter.close()
-
-		// Write header to CSV
-		dp := domesticServiceAreaPrice{}
-		csvWriter.write(dp.csvHeader())
-	}
-
-	// XLSX Sheet consts
-	const xlsxDataSheetNum int = 7  // 2a) Domestic Linehaul Prices
-	const feeColIndexStart int = 6  // start at column 6 to get the rates
-	const feeRowIndexStart int = 10 // start at row 10 to get the rates
-	const serviceAreaNumberColumn int = 2
-	const serviceAreaNameColumn int = 3
-	const serviceScheduleColumn int = 4
-	const sITPickupDeliveryScheduleColumn int = 5
-	const numEscalationYearsToProcess int = sharedNumEscalationYearsToProcess
-
-	if xlsxDataSheetNum != sheetIndex {
-		return fmt.Errorf("parseDomesticServiceAreaPrices expected to process sheet %d, but received sheetIndex %d", xlsxDataSheetNum, sheetIndex)
-	}
-
-	dataRows := params.xlsxFile.Sheets[xlsxDataSheetNum].Rows[feeRowIndexStart:]
-	for _, row := range dataRows {
-		colIndex := feeColIndexStart
-		// For number of baseline + Escalation years
-		for escalation := 0; escalation < numEscalationYearsToProcess; escalation++ {
-			// For each Rate Season
-			for _, r := range rateSeasons {
-				domPrice := domesticServiceAreaPrice{
-					ServiceAreaNumber:         getInt(getCell(row.Cells, serviceAreaNumberColumn)),
-					ServiceAreaName:           getCell(row.Cells, serviceAreaNameColumn),
-					ServiceSchedule:           getInt(getCell(row.Cells, serviceScheduleColumn)),
-					SITPickupDeliverySchedule: getInt(getCell(row.Cells, sITPickupDeliveryScheduleColumn)),
-					Season:                    r,
-					Escalation:                escalation,
-				}
-
-				domPrice.ShorthaulPrice = removeFirstDollarSign(getCell(row.Cells, colIndex))
-				colIndex++
-				domPrice.OriginDestinationPrice = removeFirstDollarSign(getCell(row.Cells, colIndex))
-				colIndex += 3 // skip 2 columns pack and unpack
-				domPrice.OriginDestinationSITFirstDayWarehouse = removeFirstDollarSign(getCell(row.Cells, colIndex))
-				colIndex++
-				domPrice.OriginDestinationSITAddlDays = removeFirstDollarSign(getCell(row.Cells, colIndex))
-				colIndex++ // skip column SIT Pickup / Delivery ≤50 miles (per cwt)
-
-				if params.showOutput == true {
-					log.Println(domPrice.toSlice())
-				}
-				if csvWriter != nil {
-					csvWriter.write(domPrice.toSlice())
-				}
-
-				colIndex += 2 // skip 1 column (empty column) before starting next Rate type
-			}
-
-		}
-	}
-
 	return nil
 }
