+++ cmd/parse_ratengine_data_ghc/main_test.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/suite"
-	"github.com/tealeg/xlsx"
 	"go.uber.org/zap"
 
 	"github.com/transcom/mymove/pkg/testingsuite"
@@ -304,64 +303,4 @@ func (suite *ParseRateEngineGHCXLSXSuite) helperTestExpectedFileOutput(goldenFil
 
 	// Remove file generated from test after compare is finished
 	os.Remove(currentOutputFilename)
-}
-
-// Test_parseDomesticLinehaulPrices
-func (suite *ParseRateEngineGHCXLSXSuite) Test_parseDomesticLinehaulPrices() {
-	initDataSheetInfo()
-	params := paramConfig{
-		processAll:   false,
-		showOutput:   false,
-		xlsxFilename: stringPointer("fixtures/pricing_template_2019-09-19_fake-data.xlsx"),
-		xlsxSheets:   []string{"6"},
-		saveToFile:   true,
-		runTime:      time.Now(),
-		runVerify:    true,
-	}
-
-	xlsxFile, err := xlsx.OpenFile(*params.xlsxFilename)
-	params.xlsxFile = xlsxFile
-	if err != nil {
-		log.Fatalf("Failed to open file %s with error %v\n", *params.xlsxFilename, err)
-	}
-
-	const sheetIndex int = 6
-	err = parseDomesticLinehaulPrices(params, sheetIndex)
-	suite.NoError(err, "parseDomesticLinehaulPrices function failed")
-
-	outputFilename := xlsxDataSheets[sheetIndex].generateOutputFilename(sheetIndex, params.runTime)
-
-	const goldenFilename string = "6_2a_domestic_linehaul_prices_golden.csv"
-	suite.helperTestExpectedFileOutput(goldenFilename, outputFilename)
-}
-
-// Test_parseDomesticServiceAreaPrices
-func (suite *ParseRateEngineGHCXLSXSuite) Test_parseDomesticServiceAreaPrices() {
-
-	initDataSheetInfo()
-	params := paramConfig{
-		processAll:   false,
-		showOutput:   false,
-		xlsxFilename: stringPointer("fixtures/pricing_template_2019-09-19_fake-data.xlsx"),
-		xlsxSheets:   []string{"7"},
-		saveToFile:   true,
-		runTime:      time.Now(),
-		runVerify:    true,
-	}
-
-	xlsxFile, err := xlsx.OpenFile(*params.xlsxFilename)
-	params.xlsxFile = xlsxFile
-	if err != nil {
-		log.Fatalf("Failed to open file %s with error %v\n", *params.xlsxFilename, err)
-	}
-
-	const sheetIndex int = 7
-	err = parseDomesticServiceAreaPrices(params, sheetIndex)
-	suite.NoError(err, "parseDomesticServiceAreaPrices function failed")
-
-	outputFilename := xlsxDataSheets[sheetIndex].generateOutputFilename(sheetIndex, params.runTime)
-
-	const goldenFilename string = "7_2b_domestic_service_area_prices_golden.csv"
-	suite.helperTestExpectedFileOutput(goldenFilename, outputFilename)
-
 }
